Early return in utils if node does not have tagName
If any node between the first and the last nodes of the scrollyteller is
followed by text, or whitespace (like a carriage return) the call to
nextSibling returns a Text node is which does not have a `tagName`
property, and does have a `querySelectorAll` method.

When `querySelectorAll` is called, in the changes introduced in commit
0a43e75 the text nodes cause an exception like this:
```
TypeError: s.querySelectorAll is not a function
```

The inability to have newlines between nodes, makes it very difficult to
configure panel node markup as all the panels have to be on the same
line.

It might be better to only reject if node does not respond to
`querySelectorAll`. I'm not sure if a Text node is valid as a
scrollyTeller panel node.